+++ example/package.json
@@ -3,38 +3,11 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "build": "./gradlew assembleDebug",
-    "test": "babel-node ./tests/utils/run.js",
-    "lint": "eslint ./**/*.js",
-    "appium": "appium"
+    "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
     "react": "15.3.2",
     "react-native": "0.34.1",
-    "tipsi-dropdown": "file:../",
-    "tape": "^4.6.0",
-    "adbkit": "^2.4.1",
-    "axios": "^0.13.1",
-    "glob": "^7.0.5",
-    "lodash": "^4.14.2",
-    "number-to-words": "^1.2.3",
-    "tap-diff": "^0.1.1",
-    "babel-cli": "^6.11.4",
-    "babel-plugin-transform-runtime": "^6.12.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.11.6",
-    "tape-async": "^2.1.1",
-    "webdriverio": "^4.1.1"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-eslint": "^6.1.2",
-    "appium": "^1.5.3",
-    "babel-eslint": "^6.1.2",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-plugin-import": "^1.13.0"
+    "tipsi-dropdown": "../"
   }
 }
