--- example/tests/utils/appium-is-running.js
@@ -1,24 +1,0 @@
-import http from 'http'
-
-export default function appiumIsRunning(host, port) {
-    return new Promise((resolve, reject) => {
-        const req = http.request({
-            method: 'GET',
-            host,
-            port,
-            path: '/wd/hub/status',
-        }, (res) => {
-            if (res.statusCode !== 200) {
-                reject(`Appium is not running on: ${host}:${port}`)
-                return
-            }
-            return resolve()
-        })
-
-        req.on('error', () => {
-            reject(`Appium is not running on: ${host}:${port}`)
-        })
-
-        req.end()
-    })
-}
