--- example/tests/utils/helper/release.js
@@ -1,7 +1,0 @@
-export default async function () {
-    if (this.driver) {
-        await this.driver.end()
-        this.driver = null
-        this.config = {}
-    }
-}
