+++ package.json
@@ -7,7 +7,7 @@
     "example": "example"
   },
   "scripts": {
-    "test": "echo \"No tests\""
+    "test": "babel-node ./test/utils/run.js"
   },
   "repository": {
     "type": "git",
@@ -23,11 +23,23 @@
     "react-native": "*"
   },
   "devDependencies": {
+    "adbkit": "^2.6.0",
+    "appium": "^1.5.3",
+    "babel-cli": "^6.16.0",
     "babel-eslint": "^7.0.0",
+    "babel-plugin-transform-runtime": "^6.15.0",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-stage-0": "^6.16.0",
+    "babel-register": "^6.16.3",
     "eslint": "^3.7.1",
     "eslint-config-airbnb": "^12.0.0",
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.3.0"
+    "eslint-plugin-react": "^6.3.0",
+    "glob": "^7.1.0",
+    "tap-diff": "^0.1.1",
+    "tape": "^4.6.2",
+    "tape-async": "^2.1.1",
+    "webdriverio": "^4.2.16"
   }
 }
