+++ test/utils/helper/index.js
@@ -1,0 +1,20 @@
+import init from './init'
+import release from './release'
+
+class Helper {
+  driver = null
+  config = {}
+
+  init = config => init.call(this, config)
+
+  release = () => release.call(this)
+
+  idFromXPath = xpath => xpath.replace(/\s+/g, '', '')
+
+  idFromResourceId = resourceId => `//*[@resource-id="${resourceId}"]`
+
+  idFromText = text => `//*[@text="${text}"]`
+}
+
+// This is singleton
+export default new Helper()
